+++ models/UserModel.php
@@ -60,8 +60,16 @@ class UserModel extends BaseModel {
      */
     public function deleteUserById($id) {
         if(is_int($id)==true||is_string($id)==true){
-            $sql = 'DELETE FROM users WHERE id = '.'$id';
-            return $this->delete($sql);      
+            $user = $this->findUserById($id);
+            if ($user!='error') {
+                $sql = 'DELETE FROM users WHERE id = '.$id;               
+                return $this->delete($sql);
+            } else {
+                return false;
+            }             
+        }
+        else{
+            return false;
         }
     }
 
@@ -92,7 +100,18 @@ class UserModel extends BaseModel {
 
         return $user;
     }
+    /**
+     * getId
+     * @param $input
+     * @return mixed
+     */
+    public function getID()
+    {
+        $sql = 'SELECT MAX(id) as user_id FROM users';
+        $user = $this->select($sql);
 
+        return $user[0]["user_id"];
+    }
     /**
      * Search users
      * @param array $params
